# openapi.yaml
openapi: 3.1.0
info:
  title: News AI Agent API
  description: "\n    News AI Agent REST API\n    \n    A production-ready API for\
    \ managing RSS feeds, articles, and triggering AI-powered workflows.\n    \n \
    \   ## Features\n    \n    - **Feed Management**: CRUD operations for RSS feed\
    \ sources\n    - **Article Management**: Query and manage news articles\n    -\
    \ **Workflows**: Trigger n8n workflows for scraping, summarization, and email\
    \ digests\n    - **System Monitoring**: Health checks and statistics\n    \n \
    \   ## Authentication\n    \n    Most endpoints require an API key in the `X-API-Key`\
    \ header.\n    The `/api/health` endpoint is public and does not require authentication.\n\
    \    \n    ## Documentation\n    \n    - **Swagger UI**: `/docs` - Interactive\
    \ API documentation\n    - **ReDoc**: `/redoc` - Alternative documentation format\n\
    \    - **OpenAPI JSON**: `/openapi.json` - Machine-readable API specification\n\
    \    "
  version: 1.0.0
paths:
...
  /api/recipients:
    post:
      tags:
      - Recipient Management
      summary: Add Recipient
      description: Add a new email recipient for news digests. Can be an individual
        email or a distribution list.
      operationId: create_recipient_api_recipients_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipientCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipient'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '400':
          description: Recipient with this email already exists
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Recipient Management
      summary: List Recipients
      description: List all digest recipients. Use enabled_only=true to filter to
        active recipients.
      operationId: list_recipients_api_recipients_get
      parameters:
      - name: enabled_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Enabled Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipient'
                title: Response List Recipients Api Recipients Get
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/workflows/scrape:
    post:
      tags:
      - Workflows
      summary: Trigger Scrape
      description: Trigger RSS feed scraping workflow.
      operationId: trigger_scrape_api_workflows_scrape_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrapeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '502':
          description: n8n webhook returned error
        '503':
          description: n8n service unavailable or timeout
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/workflows/summarize:
    post:
      tags:
      - Workflows
      summary: Trigger AI summarization
      description: "Trigger the AI summarization workflow for unprocessed articles.\n\
        \n**Limit parameter (query):**\n- `limit` not provided: processes default\
        \ 10 articles\n- `limit=0`: processes ALL unprocessed articles  \n- `limit=N`:\
        \ processes up to N articles (max 1000)\n\n**article_ids parameter (body):**\n\
        - If provided, only summarize these specific articles (ignores limit)\n- If\
        \ not provided, summarize oldest unprocessed articles up to limit"
      operationId: trigger_summarize_api_workflows_summarize_post
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 1000
            minimum: 0
          - type: 'null'
          description: Number of articles to process. 0 = all, default = 10
          title: Limit
        description: Number of articles to process. 0 = all, default = 10
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/Body_trigger_summarize_api_workflows_summarize_post'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '502':
          description: n8n webhook returned error
        '503':
          description: n8n service unavailable or timeout
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/workflows/send-digest:
    post:
      tags:
      - Workflows
      summary: Send Email Digest
      description: 'Trigger the email digest workflow.


        **Recipients:**

        - Recipients are managed via `/api/recipients` endpoints

        - All enabled recipients receive the digest by default

        - Optionally pass `recipient_ids` to send to specific recipients only


        **Force parameter:**

        - `force=false` (default): only sends if there are unsent articles

        - `force=true`: sends even if all articles were already sent'
      operationId: trigger_send_digest_api_workflows_send_digest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendDigestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Missing API key
        '403':
          description: Invalid API key
        '502':
          description: n8n webhook returned error
        '503':
          description: n8n service unavailable or timeout
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_trigger_summarize_api_workflows_summarize_post:
      properties:
        article_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Article Ids
          description: Specific article IDs to summarize (ignores limit if provided)
      type: object
      title: Body_trigger_summarize_api_workflows_summarize_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Recipient:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        enabled:
          type: boolean
          title: Enabled
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - name
      - enabled
      - created_at
      - updated_at
      title: Recipient
      description: Digest recipient response model.
    RecipientCreate:
      properties:
        email:
          type: string
          maxLength: 255
          minLength: 5
          pattern: ^[^@]+@[^@]+\.[^@]+$
          title: Email
        name:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Name
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - email
      title: RecipientCreate
      description: Create a new digest recipient.
...
    ScrapeRequest:
      properties:
        feed_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Feed Ids
      type: object
      title: ScrapeRequest
...
    SendDigestRequest:
      properties:
        recipient_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Recipient Ids
        force:
          type: boolean
          title: Force
          default: false
      type: object
      title: SendDigestRequest
      description: 'Request to trigger email digest.


        - If recipient_ids provided: send only to those recipients

        - If not provided: send to all enabled recipients from database

        - force: send even if no new articles since last digest'
...
    SuccessResponse:
      properties:
        success:
          type: boolean
          title: Success
        message:
          type: string
          title: Message
      type: object
      required:
      - success
      - message
      title: SuccessResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Get from your .env file.
servers:
- url: http://localhost:3000
  description: Local development
- url: https://api.example.com
  description: Production (update this)
security:
- APIKeyHeader: []